Tensor inheritance (#115)
* Updated inheritance of tensors.

- All ring tensors now inherit from their equivalent semiring tensors.
- All field tensors now inherit from their equivalent ring tensors.
- Added tests for Complex64 numbers.

* Added Tests

- Removed redundant transpose classes.

* Fixed Javadoc.

* Remove duplicated code.

* Added tests.

- Fixed bugs that caused failing tests.

* Added tests.

- Fixed bugs that caused failing tests.

* Added tests.

- Fixed bugs which failing tests.

* Fix for norms.

- Updated and fixed matrix and vector norms
- Changes to norms are reflected in the Condition.java for computing condition numbers.

* Added balancing decomposition

* Small fixes.

+++ src/main/java/org/flag4j/linalg/ops/dense/field_ops/DenseFieldVectorOps.java
@@ -67,9 +67,9 @@ public final class DenseFieldVectorOps {
         double sum = 0.0;
 
         for (Complex128 c : src1) {
-            double real = ((Complex128) c).re;
-            double imag = ((Complex128) c).im;
-            sum += real * real + imag * imag;
+            double real = c.re;
+            double imag = c.im;
+            sum += real*real + imag*imag;
         }
 
         return sum;
